fix[image]: add a variable for packer accelerator
Packer uses Qemu to build the image later uploaded on the tenant
and then used on the cluster nodes.
To accelerate the build, Qemu can use kvm. However, kvm is not
available on all machines, especially on virtual hardware. This
commit adds the possibility not to use kvm for the image building.